feat(completion): add `incompleteSource`
`isIncompleteSource` is used to mark whether the completion source is
`isIncomplete` for each trigger. Previously, it only depended on
`isIncomplete` in the completion result. There might be a situation
where the previous completion timed out and did not return a result,
causing the user to not be able to trigger the completion when
continuing to input characters. `isIncompleteSource` makes each trigger
`isIncomplete`, regardless of the result of the previous one.